Remove TinyMCE branding
Community request to remove the TinyMCE branding from the editor window because it interferes with typing judge comments. TinyMCE is GPL, so I think already covered by the license in the repo. Maybe in theory should add an attribution line somewhere as well?

If I'm missing something here, just deny the PR and nuke the branch.